// frontend/constants/enum.ts
export enum MatchStatus {
  UPCOMING = 'upcoming',
  LIVE = 'live',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
}

export enum BetMarket {
  MONEYLINE = 'moneyline',
  SPREAD = 'spread',
  TOTAL = 'total',
}

export enum SelectionType {
  HOME = 'home',
  AWAY = 'away',
}

export enum CurrencyType {
  COIN = 'coin',
  CASH = 'cash',
}

export enum BetStatus {
  UNSET = 'unset',
  CONFIRMING = 'confirming',
  CONFIRMED = 'confirmed',
}